refactor: Replace magic numbers with named constants and math APIs (#149)
* Use Math/MathF PI and simplify math helpers

* use math.tau

* use named constants for degree values

* refactor to use named constants and native math constants

* add casting to fix regression

* remove cast for MAX

* remove unused M_PI constant

* add xml docs

* clearer code from h5dx

* use MathF.Tau in FBOUND_PI

* remove .0

* missed those two 2.1474836E+09s

* remove cast

+++ CutTheRope/Framework/Core/RootController.cs
@@ -78,8 +78,8 @@ namespace CutTheRope.Framework.Core
             int num2 = viewTransition;
             if (num2 - 4 <= 1)
             {
-                float num = MIN(1.0, (double)((transitionDelay - (transitionTime - lastTime)) / transitionDelay));
-                if ((double)num < 0.5)
+                float num = MIN(1, (transitionDelay - (transitionTime - lastTime)) / transitionDelay);
+                if (num < 0.5f)
                 {
                     if (prevScreenImage != null)
                     {
